* src/lib/core/sequence.c: (bt_sequence_get_bar_time): * src/lib/core/song.c: (bt_song_seek_to_play_pos), (bt_song_update_play_seek_event): * src/lib/core/wire.c: (bt_wire_unlink_machines): * tests/bt-check.c: * tests/lib/core/e-sequence.c: (BT_START_TEST): * tests/lib/core/t-sequence.c: (BT_START_TEST): * tests/lib/core/t-setup.c: (BT_START_TEST): more docs in song.c, unit-test fixes after recent sequence api changes, more debug logs in song.c/sequence.c

+++ src/lib/core/wire.c
@@ -1,4 +1,4 @@
-// $Id: wire.c,v 1.83 2006-07-23 18:21:45 ensonic Exp $
+// $Id: wire.c,v 1.84 2006-07-25 20:08:27 ensonic Exp $
 /**
  * SECTION:btwire
  * @short_description: class for a connection of two #BtMachines
@@ -284,22 +284,25 @@ static void bt_wire_unlink_machines(const BtWire *self) {
   GstElement **machines=self->priv->machines;
 
   g_assert(BT_IS_WIRE(self));
-
-  GST_DEBUG("unlink machines '%s' -> '%s'",GST_OBJECT_NAME(self->priv->src->src_elem),GST_OBJECT_NAME(self->priv->dst->dst_elem));
-  /*if(machines[PART_CONVERT] && machines[PART_SCALE]) {
-    gst_element_unlink_many(self->priv->src->src_elem, machines[PART_TEE], machines[PART_GAIN], machines[PART_CONVERT], machines[PART_SCALE], self->priv->dst->dst_elem, NULL);
-  }
-  else*/
-  if(machines[PART_CONVERT]) {
-    gst_element_unlink_many(self->priv->src->src_elem, machines[PART_TEE], machines[PART_GAIN], machines[PART_CONVERT], self->priv->dst->dst_elem, NULL);
-  }
-  /*
-  else if(machines[PART_SCALE]) {
-    gst_element_unlink_many(self->priv->src->src_elem, machines[PART_TEE], machines[PART_GAIN], machines[PART_SCALE], self->priv->dst->dst_elem, NULL);
-  }
-  */
-  else {
-    gst_element_unlink_many(self->priv->src->src_elem, machines[PART_TEE], machines[PART_GAIN], self->priv->dst->dst_elem, NULL);
+  
+  // check if wire has been properly initialized
+  if(self->priv->src->src_elem && self->priv->dst->dst_elem && machines[PART_TEE] && machines[PART_GAIN]) {
+    GST_DEBUG("unlink machines '%s' -> '%s'",GST_OBJECT_NAME(self->priv->src->src_elem),GST_OBJECT_NAME(self->priv->dst->dst_elem));
+    /*if(machines[PART_CONVERT] && machines[PART_SCALE]) {
+      gst_element_unlink_many(self->priv->src->src_elem, machines[PART_TEE], machines[PART_GAIN], machines[PART_CONVERT], machines[PART_SCALE], self->priv->dst->dst_elem, NULL);
+    }
+    else*/
+    if(machines[PART_CONVERT]) {
+      gst_element_unlink_many(self->priv->src->src_elem, machines[PART_TEE], machines[PART_GAIN], machines[PART_CONVERT], self->priv->dst->dst_elem, NULL);
+    }
+    /*
+    else if(machines[PART_SCALE]) {
+      gst_element_unlink_many(self->priv->src->src_elem, machines[PART_TEE], machines[PART_GAIN], machines[PART_SCALE], self->priv->dst->dst_elem, NULL);
+    }
+    */
+    else {
+      gst_element_unlink_many(self->priv->src->src_elem, machines[PART_TEE], machines[PART_GAIN], self->priv->dst->dst_elem, NULL);
+    }
   }
   if(machines[PART_CONVERT]) {
     GST_DEBUG("  removing convert from bin, obj->ref_count=%d",G_OBJECT(machines[PART_CONVERT])->ref_count);
